requirements.txt.txt
python-telegram-bot==20.7
deep-translator
SpeechRecognition
pydub
gtts
flask
requests